Cleanup and test coverage after review
- 重複削除: registration.rs の日付ヘルパを firestore へ統一、jws::b64url を
  es256::b64url_encode へ委譲。
- 未使用 OAuthError::RequestNotSupported variant を削除。
- clippy 修正: store.rs テストモジュールを末尾へ、tpm contains、firestore の
  不要 cast、firestore_store の deref、verify.rs の lifetime 省略、mds の
  VerifiedBlob 型エイリアス + doc 整形。
- テスト追加（今回実装分の穴埋め）: request_object の FAPI window 負例
  (exp>60分/nbf境界/iat・jti欠落/aud配列)、dpop の htu query/fragment 無視と
  未来側 iat、sig::verify_rs384。es256 thumbprint を RFC 7638 既知ベクタ固定に。

clippy 15→6（残りは外部 generic-array deprecation 3 と Result<_,Response>
イディオムの result_large_err 3 のみ）。82 ユニットテスト緑。本番挙動は不変。

+++ src/firestore.rs
@@ -52,7 +52,7 @@ pub fn rfc3339(epoch_secs: u64) -> String {
 fn civil_from_days(z: i64) -> (i64, u32, u32) {
     let z = z + 719468;
     let era = if z >= 0 { z } else { z - 146096 } / 146097;
-    let doe = (z - era * 146097) as i64;
+    let doe = z - era * 146097;
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
